+++ crates/api_models/src/consts.rs
@@ -1,5 +1,5 @@
 /// Max payment intent fulfillment expiry
-pub const MAX_ORDER_FULFILLMENT_EXPIRY: u32 = 1800;
+pub const MAX_ORDER_FULFILLMENT_EXPIRY: i64 = 1800;
 
 /// Min payment intent fulfillment expiry
-pub const MIN_ORDER_FULFILLMENT_EXPIRY: u32 = 60;
+pub const MIN_ORDER_FULFILLMENT_EXPIRY: i64 = 60;
